backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
yfinance==0.2.18
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
python-dotenv==1.0.0
openai==1.3.0
pandas==2.0.3 